+++ An_Tir_Authorization/production_settings.py
@@ -76,6 +76,7 @@ MIDDLEWARE = [
     'django.middleware.csrf.CsrfViewMiddleware',
     'django.contrib.auth.middleware.AuthenticationMiddleware',
     'django.contrib.messages.middleware.MessageMiddleware',
+    'authorizations.middleware.MaintenanceLockMiddleware',
     'django.middleware.clickjacking.XFrameOptionsMiddleware',
 ]