+++ An_Tir_Authorization/settings.py
@@ -65,6 +65,7 @@ MIDDLEWARE = [
     'django.middleware.csrf.CsrfViewMiddleware',
     'django.contrib.auth.middleware.AuthenticationMiddleware',
     'django.contrib.messages.middleware.MessageMiddleware',
+    'authorizations.middleware.MaintenanceLockMiddleware',
     'django.middleware.clickjacking.XFrameOptionsMiddleware',
 ]